requirements.txt
# Core dependencies
langgraph==1.1.0
langchain-anthropic==1.4.0
langgraph-checkpoint-sqlite==3.0.3
python-dotenv==1.2.2
fastapi==0.135.1
uvicorn==0.42.0
